Updated naming of TA monitor
